19 food tracking   remaining calories (#23)
* health data is async

* caching implemented

* pre-build

* build succeeded

* linting and cleanup

* change notification duration

* increase food search

* more linting

* polish up food tracking

* fix analyze

* fix tests

* working on cache

* remove memcache

* debug logging

* wrote our own cache manager

* use custom cache

* linting

* duplicate steps

* make charts pretty

* remove double cards

* merge changes

* enhanced search

* pretty good search

* got some more ai driven functionality

+++ lib/app/modules/intro/views/intro_small.dart
@@ -337,7 +337,7 @@ class IntroSmall extends GetWidget<IntroController> {
                   ),
                   keyboardType: TextInputType.number,
                   onChanged: (value) =>
-                      controller.setDailyCalorieIntakeGoal(int.tryParse(value) ?? 0),
+                      controller.setDailyCalorieIntakeGoal(double.tryParse(value) ?? 0),
                 ),
                 const SizedBox(height: 20),
                 TextField(
@@ -347,7 +347,7 @@ class IntroSmall extends GetWidget<IntroController> {
                   ),
                   keyboardType: TextInputType.number,
                   onChanged: (value) =>
-                      controller.setDailyCaloriesBurnedGoal(int.tryParse(value) ?? 0),
+                      controller.setDailyCaloriesBurnedGoal(double.tryParse(value) ?? 0),
                 ),
                 const SizedBox(height: 20),
                 TextField(
